+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-### 2024.07.22.1
+### Unstable
 * (BREAKING) Changed the shelf backend from the default auto-select to dumbdbm (increases 
   portability and makes behavior more predictable. Fix by either selecting the appropriate dbm
   or converting the files manually to dumbdbm)
+* (BREAKING) Added Entry class used as output for database entries to support implementation of
+  default values for database entries and better in-place editing. (For most functions behavior
+  should be completely unchanged, but type will be different, calls will pass through to the 
+  underlying dictionary or list object)
 * Added behavior to pass a backend to use (e.g. dbm backends, or a dict variable if you want 
   to make an in-memory database)
 * Added backup function for backing up live databases or for the purpose of database recovery
@@ -21,6 +25,8 @@
 * Implemented bulk write locking
 * (FIX) improved thread safety when bulk writing
 * Added logging module and some debug messages
+* Added indexing
+* (Fix) Uncaught potential OverflowError in dateutil parsing in import_tools
 
 ### 2024.06.04
 * (FIX) don't crash if dateutil is missing, making it an optional dependency